# packages/brighten-api/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Brighten API",
    "description": "Unified media processing API with configurable AI backends.\n\nBrighten API provides a simple, consistent interface for AI-powered image processing operations. Configure your preferred backend providers (Replicate, fal.ai, remove.bg, etc.) and route operations through a single API.\n\n## Features\n\n- **Provider Agnostic**: Swap AI backends without changing your code\n- **Fallback Support**: Automatic failover to backup providers  \n- **Unified Format**: Consistent request/response format across all operations\n- **Base64 I/O**: Simple base64 image encoding for easy integration\n\n## Getting Started\n\n1. Clone the repository and navigate to `packages/brighten-api`\n2. Copy `config.example.yaml` to `config.yaml`\n3. Add your API keys for the providers you want to use\n4. Run `npm run dev` to start the server\n\nSee the [GitHub repository](https://github.com/phishy/brighten/tree/main/packages/brighten-api) for detailed setup instructions.",
    "version": "0.1.0",
    "license": {
      "name": "BUSL-1.1",
      "url": "https://github.com/phishy/brighten/blob/main/LICENSE"
    },
    "contact": {
      "name": "GitHub",
      "url": "https://github.com/phishy/brighten"
    }
  },
  "servers": [
    {
      "url": "http://localhost:3001",
      "description": "Local development server"
    }
  ],
...
  "paths": {
...
    "/api/v1/restore": {
      "post": {
        "summary": "Restore",
        "description": "Restore old or damaged photos using FLUX Kontext. Fixes scratches, damage, and can colorize old photos.",
        "operationId": "restore",
        "tags": [
          "Operations"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ImageInput"
              },
              "example": {
                "image": "data:image/png;base64,iVBORw0KGgo...",
                "options": {}
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Operation successful",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ImageOutput"
                }
              }
            }
          },
          "400": {
            "description": "Bad request - missing or invalid image",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                },
                "example": {
                  "error": "Image is required"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error - operation failed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                },
                "example": {
                  "error": "Provider timeout after 30000ms"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "ImageInput": {
        "type": "object",
        "required": [
          "image"
        ],
        "properties": {
          "image": {
            "type": "string",
            "description": "Base64-encoded image. Can be a data URI (recommended) or raw base64 string.",
            "example": "data:image/png;base64,iVBORw0KGgo..."
          },
          "options": {
            "type": "object",
            "description": "Operation-specific options. See individual operation docs for available options.",
            "additionalProperties": true,
            "example": {}
          }
        }
      },
      "ImageOutput": {
        "type": "object",
        "properties": {
          "image": {
            "type": "string",
            "description": "Processed image as a data URI",
            "example": "data:image/png;base64,iVBORw0KGgo..."
          },
          "metadata": {
            "type": "object",
            "description": "Information about how the operation was processed",
            "properties": {
              "provider": {
                "type": "string",
                "description": "The provider that processed the request",
                "example": "replicate"
              },
              "model": {
                "type": "string",
                "description": "The specific model used",
                "example": "cjwbw/rembg"
              },
              "processingTime": {
                "type": "number",
                "description": "Processing time in milliseconds",
                "example": 2340
              }
            }
          }
        }
      },
      "Error": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string",
            "description": "Error message describing what went wrong"
          }
        }
      }
    }
  }
}